src/django/requirements.txt
# Django and psycopg2 are included in the django base container but repeated here
# so that the analysis container will know to install them. These should be kept
# in sync with the versions in the base container.
Django==2.2.10
psycopg2-binary==2.8.4

boto3==1.9.224
django-amazon-ses==2.1.1
django-countries==5.4
django-extensions==2.2.1
django-filter==2.2.0
django-q==1.0.2
django-storages==1.7.1
django-watchman==0.18.0
djangorestframework==3.10.3
fiona==1.8.6
future==0.17.1
pycountry==19.8.18
pyuca==1.2
requests==2.22.0
us==1.0.0

# This been removed from application code but was used in old migrations, so it can't be removed
# unless those migrations are refactored to not import it.
django-localflavor==2.2
